dbc: Refactor to split dbcfile and dbc_app
